Drop the multi-agent research workflow
