+++ docs/configuration/storage.md
@@ -123,7 +123,7 @@ search:
   vector_refine_factor: 30     # Re-ranking factor for accuracy
 ```
 
-For search behavior settings (`limit`, `max_context_chars`), see [QA and Research](qa-research.md#search-settings).
+For search behavior settings (`limit`, `max_context_chars`), see [Search and Question Answering](qa.md#search-settings).
 
 - **vector_index_metric**: Distance metric for vector similarity:
   - `cosine`: Cosine similarity (default, best for most embeddings)
