--- haiku_rag_slim/haiku/rag/agents/research/__init__.py
@@ -1,7 +1,0 @@
-from haiku.rag.agents.research.dependencies import ResearchContext, ResearchDependencies
-from haiku.rag.agents.research.models import (
-    Citation,
-    IterativePlanResult,
-    ResearchReport,
-    SearchAnswer,
-)
