+++ haiku_rag_slim/haiku/rag/app.py
@@ -500,77 +500,6 @@ class HaikuRAGApp:  # pragma: no cover
             ):
                 self.console.print(renderable)
 
-    async def research(
-        self,
-        question: str,
-        filter: str | None = None,
-    ):
-        """Run research via the pydantic-graph pipeline.
-
-        Args:
-            question: The research question
-            filter: SQL WHERE clause to filter documents
-        """
-        async with HaikuRAG(
-            db_path=self.db_path,
-            config=self.config,
-            read_only=self.read_only,
-            before=self.before,
-        ) as client:
-            self.console.print("[bold cyan]Starting research[/bold cyan]")
-            self.console.print(f"[bold blue]Question:[/bold blue] {question}")
-            self.console.print()
-
-            report = await client.research(question=question, filter=filter)
-
-            if report is None:
-                self.console.print("[red]Research did not produce a report.[/red]")
-                return
-
-            # Display the report
-            self.console.print("[bold green]Research Report[/bold green]")
-            self.console.rule()
-
-            # Title and Executive Summary
-            self.console.print(f"[bold]{report.title}[/bold]")
-            self.console.print()
-            self.console.print("[bold cyan]Executive Summary:[/bold cyan]")
-            self.console.print(report.executive_summary)
-            self.console.print()
-
-            # Main Findings
-            if report.main_findings:
-                self.console.print("[bold cyan]Main Findings:[/bold cyan]")
-                for finding in report.main_findings:
-                    self.console.print(f"• {finding}")
-                self.console.print()
-
-            # Conclusions
-            if report.conclusions:
-                self.console.print("[bold cyan]Conclusions:[/bold cyan]")
-                for conclusion in report.conclusions:
-                    self.console.print(f"• {conclusion}")
-                self.console.print()
-
-            # Recommendations
-            if report.recommendations:
-                self.console.print("[bold cyan]Recommendations:[/bold cyan]")
-                for rec in report.recommendations:
-                    self.console.print(f"• {rec}")
-                self.console.print()
-
-            # Limitations
-            if report.limitations:
-                self.console.print("[bold yellow]Limitations:[/bold yellow]")
-                for limitation in report.limitations:
-                    self.console.print(f"• {limitation}")
-                self.console.print()
-
-            # Sources Summary
-            if report.sources_summary:
-                self.console.print("[bold cyan]Sources:[/bold cyan]")
-                self.console.print(report.sources_summary)
-
     async def rebuild(self, mode: RebuildMode = RebuildMode.FULL):
         async with HaikuRAG(
             db_path=self.db_path,
