+++ haiku_rag_slim/haiku/rag/client/agents.py
@@ -1,9 +1,9 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from haiku.rag.agents.research.models import Citation, ResearchReport
     from haiku.rag.client import HaikuRAG
     from haiku.rag.sandbox import AnalysisResult
+    from haiku.rag.store.models.citation import Citation
 
 
 async def ask(
@@ -35,39 +35,6 @@ async def ask(
         if cid in state.citation_index
     ]
     return answer, citations
-
-
-async def research(
-    client: "HaikuRAG",
-    question: str,
-    *,
-    filter: str | None = None,
-    max_iterations: int | None = None,
-) -> "ResearchReport":
-    """Run multi-agent research to investigate a question.
-
-    Args:
-        client: The HaikuRAG client.
-        question: The research question to investigate.
-        filter: SQL WHERE clause to filter documents.
-        max_iterations: Override max iterations (None uses config default).
-
-    Returns:
-        ResearchReport with structured findings.
-    """
-    from haiku.rag.agents.research.dependencies import ResearchContext
-    from haiku.rag.agents.research.graph import build_research_graph
-    from haiku.rag.agents.research.state import ResearchDeps, ResearchState
-
-    graph = build_research_graph(config=client._config)
-    context = ResearchContext(original_question=question)
-    state = ResearchState.from_config(
-        context=context, config=client._config, max_iterations=max_iterations
-    )
-    state.search_filter = filter
-    deps = ResearchDeps(client=client)
-
-    return await graph.run(state=state, deps=deps)
 
 
 async def analyze(
